pyproject.toml
[project]
name = "gcse-chinese-trainer"
version = "0.1.0"
description = "GCSE Chinese Trainer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml>=0.12.39",
    "libsql>=0.1.11",
    "libsql-client>=0.3.1",
    "psycopg2-binary>=2.9.11",
    "fastsql>=2.1.1",
    "bcrypt>=5.0.0",
    "google-generativeai>=0.8.6",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
